Moved helper functions into helper.js
